Define: Improved getModuleById(), so it wont throw an exception anymore in case the module is not found. Added if (module != null) clauses in all using methods. General: Improved a number of test classes.

+++ src/husacct/define/DefineSarServiceImpl.java
@@ -56,7 +56,9 @@ public class DefineSarServiceImpl implements IDefineSarService {
 	@Override
 	public void removeModule(String logicalPath) {
 		try {
-			moduleService.removeModuleById(moduleService.getModuleByLogicalPath(logicalPath).getId());
+			if (moduleService.getModuleByLogicalPath(logicalPath) != null) {
+				moduleService.removeModuleById(moduleService.getModuleByLogicalPath(logicalPath).getId());
+			}
         } catch (Exception e) {
 	        this.logger.warn(" Exception: "  + e );
         }
